--- a/src/caffe/proto/caffe.proto
+++ b/src/caffe/proto/caffe.proto
@@ -518,6 +518,13 @@ message ConvolutionParameter {
   repeated uint32 pad = 3; // The padding size; defaults to 0
   repeated uint32 kernel_size = 4; // The kernel size
   repeated uint32 stride = 6; // The stride; defaults to 1
+  // Properties of dilated convolution are described in
+  // F. Yu et al, Multi-Scale Context Aggregation by Dilated Convolutions, 2015.
+  // "dilation" was first proposed and used in a trous algorithm, as decribed in
+  // Holschneider et al. A Real-Time Algorithm for Signal Analysis
+  // with the Help of the Wavelet Transform, 1987.
+  // It was called filter rarefaction in Long* and Shelhamer* and Darrell,
+  // Fully Convolutional Networks for Semantic Segmentation, 2014.
   repeated uint32 dilation = 18; // The dilation; defaults to 1
 
   // For 2D convolution only, the *_h and *_w versions may also be used to