… src/caffe/test/test_convolution_layer.cpp
… src/caffe/test/test_im2col_kernel.cu
… src/caffe/test/test_im2col_layer.cpp
… src/caffe/util/im2col.cu