This PR extends im2col to support dilated convolution, as described in the paper [Multi-Scale Context Aggregation by Dilated Convolutions](http://arxiv.org/abs/1511.07122). The changes aim to support general combinations of dilation and stride in convolution layer. Although it is hard to tweak cuDNN to support dilation, im2col seems a natural place to start with. Both 2D and ND im2col are changed to support dilation. The test code for im2col layer is changed to test dilation. Let me know if there is any other case that may require testing.

In addition, I found that the original implementation of ND im2col and col2im doesn't make use of CUDA memory efficiently. To improve that, I use shared memory in CUDA to cache global memory.

I ran benchmarks on alexnet to compare the performance between master and this branch. The results are:

master 2D convolution:
```
Average Forward pass: 352.594 ms.
Average Backward pass: 254.317 ms.
Average Forward-Backward: 607.01 ms.
```
dilation 2D convolution:
```
Average Forward pass: 360.493 ms.
Average Backward pass: 255.279 ms.
Average Forward-Backward: 615.875 ms.
```

master ND convolution:
```
Average Forward pass: 505.18 ms.
Average Backward pass: 398.236 ms.
Average Forward-Backward: 903.509 ms.
```

dilation ND convolution:
```
Average Forward pass: 425.503 ms.
Average Backward pass: 323.731 ms.
Average Forward-Backward: 749.334 ms.
```

I ran 50 iterations with batch size 256. The numbers show it adds small overhead to support dilation. It improves the ND convolution a lot to use shared memory.

Benchmark model for 2D: [alexnet.prototxt](https://gist.github.com/fyu/f0214efcacf7c1688c38)
Benchmark mdoel for ND: [alexnet_nd.prototxt](https://gist.github.com/fyu/3ae109c9cb671acbc7cf)