@longjon 2015-12-15
This looks pretty nice, thanks @fyu. A few initial comments:

1. There is already a similar PR for this functionality in #3232. However, it looks like this one is in better shape (e.g., #3232 is failing Travis and has neither a clean history nor a clean diff), so I'd like to proceed from here. To acknowledge the earlier implementations, I propose adding the following text to the initial commit message:

    > An early implementation of this functionality for Caffe was written by @gpapan, which was extended and improved by @tamakoji in a previous implementation of this and the next commit.

    Does that sound accurate and appropriate to everyone involved? @shelhamer, does that sound compatible with how we've handled credit distribution in similar situations in the past?

2. I believe @shelhamer has written more exhaustive tests for this functionality (e.g., upgrading the reference implementation), which should be included here. @shelhamer, perhaps you can point @fyu to relevant commits to cherry-pick?

3. @shelhamer and I discussed earlier the issue of what to call the new parameter here; at that time we did not like any of the options we could think of: `hole` or `hole_stride` is misleading as pointed out by @tamakoji; `kernel_stride` is easily confused with `stride`; and other possibilities are unclear or obscure. `dilation` seems like the best term I've heard yet, so I don't mind keeping this as-is. Opinions?

4. The perf difference for 2D im2col seems acceptable to me; do others agree? @fyu, can you explain what the source of the small difference is?

I'll make additional comments inline.