@fyu 2015-12-15
Thank @longjon and @shelhamer for the fine job to manage the caffe code. I believe the performance difference is due to the some additional operation for dilation.  Because the original code is already very concise, the difference becomes noticeable by 1% to 2%. For ND case, I guess the difference would be harder to notice if we removed the dilation, since the original implementation is more complicated than the 2D case.