@longjon 2015-12-15
Re: 2 above, it looks like the reference implementation is actually upgraded here; so @shelhamer perhaps you can check whether your tests do or do not provide additional functionality.

Additionally, the tests feel a little lackluster to me for the following reasons:

1. I don't see dilated convolution being checked against the reference implementation, as the conv tests use dilation 1.
2. The im2col tests are modified to use non-one dilation. I'd rather see _additional_ tests for non-one dilation; the current changes make me nervous as we're departing from the common case.

Other than that this looks pretty ready except as noted (all the notes being nitpicks). I haven't checked the logic in detail, but @shelhamer can comment on whether it all looks right to him, as he's wrestled with it before.

Re: the performance difference (in 2d, I'm not going to worry about nd for now), it looks like most of the gap is in `im2col`, which I find it somewhat surprising that a few 32-bit multiplies would make a noticeable difference, but I guess it does. We can always consider a separate kernel or a templated kernel to improve performance if it becomes an issue in the future (which seems unlikely).