@tamakoji 2015-12-15
@fyu @longjon @shelhamer Thanks. Yes, 'dilation' sounds more adequate than 'kernel_stride'