@fyu 2015-12-16
Thank @gpapan for the input. I guess we all agree that holes and a trous may not be a proper term since it refers to the algorithm with a pre-determined filter. We propose to name it "Dilated Convolution" because of the similarity to the traditional dilation. It will also be easier to communicate if we can keep the name a single word. If we call it input_stride, we may have to change the current "stride" to output_stride to disambiguate and it will require people to change the concept of stride, which doesn't sound ideal. So I think dilation is better to make the name short, distinct and relevant. It is definitely great to know that TensorFlow will also support this operation.