@longjon 2015-12-17
Okay, a few observations regarding naming and such. There are _three_ ways you can "add holes" in convolution:

1. You can _implicitly add holes to the kernel_. That's the subject of this patch.
2. You can _implicitly add holes to the input_. (Note that there's nothing special distinguishing input from kernel.) That's supported in `master` using the `stride` parameter of the `Deconvolution` layer.
3. You can _implicitly add holes to the output_. (If that's confusing, think of the gradient computation, which is the same as the previous case.) That's supported in `master` using the `stride` parameter of the `Convolution` layer.

Now, given the above it does seem natural to refer to the spacing parameter for case 1 as a "stride". However, `input_stride` seems particularly unfortunate for these reasons:

1. It's not case 2 above.
2. It's also reasonably confusable with case 3 above, which could be described as "the stride of the kernel in the input".
3. In the future, we may support simultaneous specification of all three spacing parameters, and in that case it seems clear that using `input_stride` for 1 breaks the symmetry of the terms.

Also keep in mind that memory layout could be strided (in a parametrized way) in the future (as is supported by cuDNN), which is another use of stride with a different meaning.

So even though it's somewhat natural to call this new parameter `*_stride`, I haven't heard a compelling choice for `*`, and given the current situation where the other two hole-making options are actually different layers, it also seems natural to pick a distinct name to avoid creating a false correspondence.

`dilation`, while a newly invented term, is at least evocative and not easily confused with regular old `stride`...