@fyu 2015-12-17
Thanks to @longjon's efforts to scrutinize the code, the format issues have been cleared out. I also add more tests in convolution to test dilation for both forward and backward passes. The tests for im2col are also changed so that the special cases for dilation are tested separately. @shelhamer, do you have any further comments on the tests?